sleep to avoid race condition when encrypting key
we tee the random key data into two simultaneous processes to avoid
landing the unencrypted key to the filesystem. This caused a
race condition as the key encryption was not gauranteed to finish prior
to trying to combine the encrypted data into the final output.

touching a file when done, and checking for the existence of that file,
should close the race. This is a quick&dirty fix, but if you weren't
okay with quick&dirty, you wouldn't be wrapping openssl commands in a
bash script anyway.